requirements.txt
playwright
requests
beautifulsoup4
python-dotenv
lxml
